14: Recursive page generation.

+++ src/copy_overwrite.py
@@ -1,15 +1,14 @@
 import os
 import shutil
+import re
 
 def replace_with_copy(source, destination, delete=False):
-    src = os.path.abspath(source)
-    dst = os.path.abspath(destination)
     if delete == True:
         shutil.rmtree(destination)
         os.mkdir(destination)
     for entity in os.listdir(source):
         ent = os.path.join(source, entity)
-        if os.path.isfile(ent):
+        if os.path.isfile(ent) and not re.findall(r".+.md(?!.)", entity):
             shutil.copy(ent, destination)
         if os.path.isdir(ent):
             copy_ent = os.path.join(destination, entity)
